[SIGNED-OFF] Bug 16190: Enable the indexer daemon by default
This patch enables the indexer daemon by default on packages installs.
It does so by setting USE_INDEXER_DAEMON=yes on the /etc/default/koha-common
file. It does remove now irrelevant comments from that file.

This patch is straightforward to review. If is just a matter of choosing
to enable or not to enable by default.

Kind regards

PS: Let end users rejoice this so far hidden feature :-D

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher <bredan@bywatersolutions.com>