Bug 9387 - Feedback message for FAILED check out items are not obvious for visually impaired
This patch adds a heading to the error dialg displayed when a barcode
has been submitted for checkout which doesn't exist. This patch also
modifies the markup of the error dialog in order to make error a little
more clear to sighted users.

To test, apply the patch and submit a barcode for checkout which does
not exist. The resulting error dialog should have a heading, "Barcode
not found."

Signed-off-by: Marc Veron <veron@veron.ch>
Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>