Bug 16341: Revise the way table controls look on the title detail page
This patch makes changes to the way links are displayed on the
bibliographic detail page, adding Font Awesome icons to make the links
clearer. Some instances of the "onclick" attribute are removed.

This patch also converts the "Edit" link to a Bootstrap button.

To test, apply the patch and locate a title in the catalog with multiple
holdings from different branches.

Test all the table controls under a variety of conditions:

- Logged in as a user who can or can't edit items
- Logged in as a user who can or can't perform batch item operations.
- With the StaffDetailItemSelection system preference enabled or
  disabled.
- With the SeparateHoldings system preference turned on or off.

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as advertised.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>