Bug 16360: Adding 'actions' class to column in tag review so buttons don't wrap
EDIT: Comment 3

Note: the wrapping only happens when all tags have been approved (or all
rejected), i.e. column has a certain width. When testing, ensure all
tags have been approved or rejected and click the opposite
action to see the effect of the bug and the patch.

To test:
1) Go to Tools -> Tags
2) Click Reject for a tag
3) Notice buttons do not wrap

Sponsored-by: Catalyst IT
Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>