Bug 21803: Redesign authorized values interface
This patch redesigns the authorized values management interface. The
primary change is that the initial view of authorized values is now a
table listing all authorized values categories and their descriptions.
Descriptions have been added or updated according to the text of the
manual.

To test, apply the patch and go to Administration -> Authorized values.

 - You should see a table of authorized values displayed in a
   DataTables-formatted table.
   - Click a category. You should be taken to the list of authorized
     values for that category. If there are none defined you should see
     a message saying so, including a button to add one.
   - Click the "Add" button next to an authorized value category. It
     should take you to the form for adding an authorized value to that
     category.

Signed-off-by: Pierre-Marc Thibault <pierre-marc.thibault@inLibro.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>