Bug 16566 [Revised] 'Print slip' button formatting inconsistent
This patch modifies the style of some "print slip" links on the checkin
page to make them consistent with other buttons shown in dialogs.

This patch also removes the use of "onclick" in some places in favor of
defining events in the JavaScript.

To test apply the patch and go to Circulation -> Check in.

- Check in an item which belongs at another branch. You should see a
  dialog which asks the user to transfer the item. Confirm that the
  "Print slip" button looks correct and works correctly.

- Check in the same item again. You should see a different dialog which
  has both a "Print transfer slip" and "Cancel transfer" button. The
  "Print transfer slip" button should look correct and work correctly.

- Check in an item which is part of a rotating collection. (Or, if Bug
  16570 is still open, check in any item) You should see a dialog which
  notifies you that the item is part of a rotating collection. The
  "Print slip" button in this dialog should look correct and work
  correctly.

Followed test plan. Buttons appear as expected.
Signed-off-by: Marc Véron <veron@veron.ch>

Revision following QA: Removed redundant class attribute from the print
button tested in step 1.

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>