Bug 16762: Record matching rules: Remove match check link removes too much
My patch for Bug 5006 introduced a lot of markup changes to the record
matching rules template. I missed a change in one are though, which
caused a bug with the way the 'Remove match check' link works.

To reproduce:

- Go to Administration -> Record matching rules.
- Edit a matching rule which includes at least one match check.
- Click a 'Remove match check' link. The entire "Required match checks"
  block is removed.

To test, apply the patch and repeat the steps above. Clicking any
'Remove match check' link should remove only the corresponding rule.

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as advertised

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>